+++ unlock-js/src/v01/index.js
@@ -2,6 +2,7 @@ import abis from '../abis'
 import createLock from './createLock'
 import getLock from './getLock'
 import purchaseKey from './purchaseKey'
+import grantKey from './grantKey'
 import updateKeyPrice from './updateKeyPrice'
 import withdrawFromLock from './withdrawFromLock'
 import configureUnlock from './configureUnlock'
@@ -12,6 +13,7 @@ export default {
   configureUnlock,
   getLock,
   purchaseKey,
+  grantKey,
   updateKeyPrice,
   withdrawFromLock,
   isLockManager,